app/requirements.txt
psycopg2 == 2.9.9
biopython == 1.83
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
biopython==1.83
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
django-environ==0.11.2
kombu==5.3.5
numpy==1.26.4
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
redis==5.0.3
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
